Document payment type codes
